make pip install-able
Added scaffolding so that `pip install` works. Driving force was so
Windows user doesn't have to create a batch file to run the script.

Only functional change from upstream master is that the script is called
"mutagenc" instead of mutagen. (In order to avoid name collision. Didn't
use mutagen-cli because harder to type. Borrowed "c" suffix convention
for CLI from Ghostscript.)

This is the first time I've done this for another program. I may have
not done it correctly.